Moved ScanQRCodeFromFile from Activities to Utilities
